+++ src/models/jsonIndex.ts
@@ -2,12 +2,16 @@
 // Elements are sorted and should be displayed in the order they are added to the index.
 export type JsonIndex = JsonIndexEntry[]
 
+export type IndexId = string & { __brand__: "IndexId" }
+
 export interface JsonIndexEntry {
-	id: string
+	id: IndexId
+	parentId?: IndexId
 	type: JsonIndexTypes
 	file: string
 }
 
 export enum JsonIndexTypes {
 	chest = "chest",
+	chestAddon = "chest_addon",
 }